# Conflicts:
#   frontend/src/pages/Diary/Diary.scss

+++ backend/src/main/java/com/ieng/ieng/domain/diary/service/DiaryServiceImpl.java
@@ -10,20 +10,20 @@ import com.ieng.ieng.domain.diary.repository.DiaryKeywordRepository;
 import com.ieng.ieng.domain.diary.repository.DiaryRepository;
 import com.ieng.ieng.domain.member.entity.Member;
 import com.ieng.ieng.domain.member.repository.MemberRepository;
-import com.ieng.ieng.global.exception.DuplicateDiaryException;
-import com.ieng.ieng.global.exception.EmptyFileException;
-import com.ieng.ieng.global.exception.ExistDiaryException;
-import com.ieng.ieng.global.exception.NoExistMemberException;
+import com.ieng.ieng.global.exception.*;
 import com.ieng.ieng.global.s3.S3UploaderServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+
+import org.hibernate.PropertyValueException;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.dao.IncorrectResultSizeDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
@@ -86,23 +86,34 @@ public class DiaryServiceImpl implements DiaryService{
 
 
     @Override
-    public void createDiary(String email, DiaryRequestDto diaryRequestDto){
+    public void createDiary(String email, DiaryRequestDto diaryRequestDto, MultipartFile multipartFile){
 
         Member member = memberRepository.findByEmail(email).orElseThrow(() -> new NoExistMemberException("존재하는 회원정보가 없습니다."));
+        Diary diary = null;
 
         LocalDate date = LocalDate.now();
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+
         if(diaryRepository.existsByMemberAndDiaryDTTM(member,date.format(formatter))){
             throw new ExistDiaryException("오늘 이미 일기 작성 완료.");
         }
-        Diary diary = Diary.builder()
+
+
+        Diary diary2 = diaryRepository.findDiaryByMemberAndDiaryDTTM(member, date.format(formatter));
+
+        if(multipartFile.isEmpty()){
+            throw new EmptyFileException("사진 파일은 필수 값입니다.");
+        }
+
+        diary = Diary.builder()
                 .member(member)
-                .diaryPicturePath(diaryRequestDto.getPicturePath())
                 .diaryContent(diaryRequestDto.getContent())
                 .diaryEmotion(diaryRequestDto.getEmotion())
                 .diaryDTTM(date.format(formatter))
                 .build();
         diaryRepository.save(diary);
+
+
         List<DiaryKeywordDto> diaryKeywordList = diaryRequestDto.getDiaryKeywordList();
         for(DiaryKeywordDto diaryKeywordDto : diaryKeywordList){
             DiaryKeyword diaryKeyword = DiaryKeyword.builder()
@@ -115,10 +126,12 @@ public class DiaryServiceImpl implements DiaryService{
     }
     @Override
     public void uploadDiaryImage(MultipartFile multipartFile, String email) {
+        Member member = memberRepository.findByEmail(email).orElseThrow(() -> new NoExistMemberException("존재하는 회원정보가 없습니다."));
+
         if(multipartFile.isEmpty()){
             throw new EmptyFileException("파일이 없습니다.");
         }
-        Member member = memberRepository.findByEmail(email).orElseThrow(() -> new NoExistMemberException("존재하는 회원정보가 없습니다."));
+
 
         LocalDate date = LocalDate.now();
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
@@ -127,10 +140,10 @@ public class DiaryServiceImpl implements DiaryService{
         int index = multipartFile.getOriginalFilename().indexOf(".");
         String fileNameExtension = multipartFile.getOriginalFilename().substring(index);
         UUID uuid = UUID.randomUUID();
-        String fileName= "user/"+email+"/diary"+"/"+today+"/"+uuid+fileNameExtension;
+        String fileName = "user/" + email + "/diary" + "/" + today + "/" + uuid + fileNameExtension;
 
-        Diary diary = diaryRepository.findDiaryByMemberAndDiaryDTTM(member,today);
-        diary.updateDiaryPicturePath(s3Domain+fileName);
+        Diary diary = diaryRepository.findDiaryByMemberAndDiaryDTTM(member, today);
+        diary.updateDiaryPicturePath(s3Domain + fileName);
         diaryRepository.save(diary);
         s3UploaderService.uploadPicture(multipartFile, fileName);
 
@@ -138,8 +151,15 @@ public class DiaryServiceImpl implements DiaryService{
     @Override
     public void deleteDiary(String email, DiaryDeleteDto diaryDeleteDto){
         Member member = memberRepository.findByEmail(email).orElseThrow(() -> new NoExistMemberException("존재하는 회원정보가 없습니다."));
-        Diary diary = diaryRepository.findDiaryByMemberAndDiaryDTTM(member, diaryDeleteDto.getDate());
-        diaryRepository.delete((diary));
+
+        if (diaryRepository.existsByMemberAndDiaryDTTM(member, diaryDeleteDto.getDate())){
+            Diary diary = diaryRepository.findDiaryByMemberAndDiaryDTTM(member, diaryDeleteDto.getDate());
+            diaryRepository.delete((diary));
+        }
+        else{
+            throw new EmptyDiaryException("해당날짜에 조회되는 일기가 없습니다.");
+        }
+
     }
 
 }
